Добавил базовую поддержку создания заявок Впереди тесты и остальные методы

+++ domain/src/main/java/by/mmb/model/transportationRequest/Cargo.java
@@ -26,7 +26,7 @@ public class Cargo {
     /**
      * вес
      */
-    private double weight;
+    private int weight;
     /**
      * длина
      */
